\name{pd.atdschip.tiling}
\Rdversion{1.1}
\docType{package}
\alias{pd.atdschip.tiling}
\title{
  Annotation package for pd.atdschip.tiling.
}
\description{
  Annotation package for pd.atdschip.tiling built with pdInfoBuilder.
}
\details{
  This package is to be used in conjunction with the oligo package.
}
\keyword{package}
